Move DetailsWidgetUser to core components
