+++ src/modules/core/components/DetailsWidgetUser/DetailsWidgetUser.css
@@ -6,8 +6,8 @@
 .address {
   margin-left: 8px;
   font-size: var(--size-smallish);
-  line-height: 1;
   font-weight: var(--weight-bold);
+  line-height: 1;
   color: var(--temp-grey-blue-7);
   cursor: pointer;
 }
@@ -15,7 +15,7 @@
 .username {
   margin-left: 8px;
   font-size: var(--size-normal);
-  line-height: 1;
   font-weight: var(--weight-bold);
+  line-height: 1;
   color: var(--pink);
 }
