+++ src/modules/core/components/DetailsWidgetUser/DetailsWidgetUser.css.d.ts
