+++ src/modules/core/components/DetailsWidgetUser/index.ts
