+++ src/modules/dashboard/components/ActionsPage/DetailsWidget/index.ts
@@ -1,4 +1,3 @@
 export { default } from './DetailsWidget';
 export * from './DetailsWidget';
-export { default as DetailsWidgetUser } from './DetailsWidgetUser';
 export { default as DetailsWidgetTeam } from './DetailsWidgetTeam';
